refactor: Comply with Wdouble-promotion
Remove implicit float to double conversions.

+++ src/core/toxfileprogress.cpp
@@ -129,8 +129,8 @@ double ToxFileProgress::getTimeLeftSeconds() const
     }
 
     const auto speed = getSpeed();
-    if (speed == 0.0f) {
-        return std::numeric_limits<float>::infinity();
+    if (speed == 0.0) {
+        return std::numeric_limits<double>::infinity();
     }
 
     return double(filesize - samples[activeSample].bytesSent) / getSpeed();
